// Corte II/Projects/03-DynamicApp/src/models/recipe.model.ts
export interface IRecipeResponse {
    recipes: IRecipe[];
    total:   number;
    skip:    number;
    limit:   number;
}

export interface IRecipe {
    id:                 number;
    name:               string;
    ingredients:        string[];
    instructions:       string[];
    prepTimeMinutes:    number;
    cookTimeMinutes:    number;
    servings:           number;
    difficulty:         Difficulty;
    cuisine:            string;
    caloriesPerServing: number;
    tags:               string[];
    userId:             number;
    image:              string;
    rating:             number;
    reviewCount:        number;
    mealType:           string[];
}

export type Difficulty = "Easy" | "Medium";